Revert redundant client module renames
Rename event_bus_client.py and exception_trap_client.py back to client.py.
The original client.py naming was appropriate as the directory structure
already provided clear context (da_vinci.event_bus.client and
da_vinci.exception_trap.client). The verbose naming added no additional
clarity and was inconsistent with Python naming conventions.

Updated all imports across both core and cdk packages to reflect the
reverted naming.
