+++ packages/cdk/da_vinci_cdk/framework_stacks/services/event_bus/runtime/bus.py
@@ -12,8 +12,8 @@ from uuid import uuid4
 import boto3
 
 from da_vinci.core.json import DaVinciObjectEncoder
+from da_vinci.event_bus.client import EventResponder, EventResponseStatus
 from da_vinci.event_bus.event import Event
-from da_vinci.event_bus.event_bus_client import EventResponder, EventResponseStatus
 from da_vinci.event_bus.tables.event_bus_subscriptions import EventBusSubscriptions
